+++ Frends.LinkedIn.SearchAdCampaigns/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [1.0.1] - 2023-10-31
+### Added
+- Added multitagreting to net471 and netstandard2.0
+
 ## [1.0.0] - 2023-10-05
 ### Changed
 - Initial implementation
