+++ Frends.LinkedIn.SearchAdCampaigns/Frends.LinkedIn.SearchAdCampaigns/Definitions/CampaignStatus.cs
@@ -1,18 +1,19 @@
-﻿namespace Frends.LinkedIn.SearchAdCampaigns.Definitions;
-
-/// <summary>
-/// Campaign status filters.
-/// </summary>
-public enum CampaignStatus
+﻿namespace Frends.LinkedIn.SearchAdCampaigns.Definitions
 {
+    /// <summary>
+    /// Campaign status filters.
+    /// </summary>
+    public enum CampaignStatus
+    {
 #pragma warning disable CS1591 // self-explanatory
-    ACTIVE,
-    PAUSED,
-    ARCHIVED,
-    COMPLETED,
-    CANCELED,
-    DRAFT,
-    PENDING_DELETION,
-    REMOVED,
+        ACTIVE,
+        PAUSED,
+        ARCHIVED,
+        COMPLETED,
+        CANCELED,
+        DRAFT,
+        PENDING_DELETION,
+        REMOVED,
 #pragma warning restore CS1591 // self-explanatory
+    }
 }
