+++ Frends.LinkedIn.SearchAdCampaigns/Frends.LinkedIn.SearchAdCampaigns/Definitions/CampaignType.cs
@@ -1,14 +1,15 @@
-﻿namespace Frends.LinkedIn.SearchAdCampaigns.Definitions;
-
-/// <summary>
-/// Campaign type filters.
-/// </summary>
-public enum CampaignType
+﻿namespace Frends.LinkedIn.SearchAdCampaigns.Definitions
 {
+    /// <summary>
+    /// Campaign type filters.
+    /// </summary>
+    public enum CampaignType
+    {
 #pragma warning disable CS1591 // self-explanatory
-    TEXT_AD,
-    SPONSORED_UPDATES,
-    SPONSORED_INMAILS,
-    DYNAMIC,
+        TEXT_AD,
+        SPONSORED_UPDATES,
+        SPONSORED_INMAILS,
+        DYNAMIC,
 #pragma warning restore CS1591 // self-explanatory
+    }
 }
