+++ Frends.LinkedIn.SearchAdCampaigns/Frends.LinkedIn.SearchAdCampaigns/Definitions/Filter.cs
@@ -1,49 +1,50 @@
-﻿namespace Frends.LinkedIn.SearchAdCampaigns.Definitions;
-
-using System.ComponentModel;
-using System.ComponentModel.DataAnnotations;
-
-/// <summary>
-/// Input class usually contains parameters that are required.
-/// </summary>
-public class Filter
+﻿namespace Frends.LinkedIn.SearchAdCampaigns.Definitions
 {
-    /// <summary>
-    /// Ad Account Id.
-    /// </summary>
-    /// <example>506333826</example>
-    [DefaultValue("")]
-    [DisplayFormat(DataFormatString = "Text")]
-    public string AdAccountId { get; set; }
+    using System.ComponentModel;
+    using System.ComponentModel.DataAnnotations;
 
     /// <summary>
-    /// Searches for accounts by reference.
+    /// Input class usually contains parameters that are required.
     /// </summary>
-    [DefaultValue("")]
-    [DisplayFormat(DataFormatString = "Text")]
-    public string Reference { get; set; }
+    public class Filter
+    {
+        /// <summary>
+        /// Ad Account Id.
+        /// </summary>
+        /// <example>506333826</example>
+        [DefaultValue("")]
+        [DisplayFormat(DataFormatString = "Text")]
+        public string AdAccountId { get; set; }
 
-    /// <summary>
-    /// Searches for an account by name.
-    /// </summary>
-    [DefaultValue("")]
-    [DisplayFormat(DataFormatString = "Text")]
-    public string Name { get; set; }
+        /// <summary>
+        /// Searches for accounts by reference.
+        /// </summary>
+        [DefaultValue("")]
+        [DisplayFormat(DataFormatString = "Text")]
+        public string Reference { get; set; }
 
-    /// <summary>
-    /// Searches for an account by ID.
-    /// </summary>
-    [DefaultValue("")]
-    [DisplayFormat(DataFormatString = "Text")]
-    public string Id { get; set; }
+        /// <summary>
+        /// Searches for an account by name.
+        /// </summary>
+        [DefaultValue("")]
+        [DisplayFormat(DataFormatString = "Text")]
+        public string Name { get; set; }
 
-    /// <summary>
-    /// Filters for LInkedIn request.
-    /// </summary>
-    public TypeFilter[] TypeFilters { get; set; }
+        /// <summary>
+        /// Searches for an account by ID.
+        /// </summary>
+        [DefaultValue("")]
+        [DisplayFormat(DataFormatString = "Text")]
+        public string Id { get; set; }
 
-    /// <summary>
-    /// Filters for LInkedIn request.
-    /// </summary>
-    public StatusFilter[] StatusFilters { get; set; }
+        /// <summary>
+        /// Filters for LInkedIn request.
+        /// </summary>
+        public TypeFilter[] TypeFilters { get; set; }
+
+        /// <summary>
+        /// Filters for LInkedIn request.
+        /// </summary>
+        public StatusFilter[] StatusFilters { get; set; }
+    }
 }
