+++ Frends.LinkedIn.SearchAdCampaigns/Frends.LinkedIn.SearchAdCampaigns/Definitions/TypeFilter.cs
@@ -1,13 +1,14 @@
-﻿namespace Frends.LinkedIn.SearchAdCampaigns.Definitions;
-
-/// <summary>
-/// Filter object used with LinkedIn request.
-/// </summary>
-public class TypeFilter
+﻿namespace Frends.LinkedIn.SearchAdCampaigns.Definitions
 {
     /// <summary>
-    /// Campaign type filter.
+    /// Filter object used with LinkedIn request.
     /// </summary>
-    /// <example>CampaignType.TEXT_AD</example>
-    public CampaignType CampaignType { get; set; }
+    public class TypeFilter
+    {
+        /// <summary>
+        /// Campaign type filter.
+        /// </summary>
+        /// <example>CampaignType.TEXT_AD</example>
+        public CampaignType CampaignType { get; set; }
+    }
 }
