+++ Frends.LinkedIn.SearchAdCampaigns/Frends.LinkedIn.SearchAdCampaigns/Extensions.cs
@@ -1,26 +1,27 @@
-﻿namespace Frends.LinkedIn.SearchAdCampaigns;
-
-using System;
-using System.Net.Http;
-using Frends.LinkedIn.SearchAdCampaigns.Definitions;
-
-internal static class Extensions
+﻿namespace Frends.LinkedIn.SearchAdCampaigns
 {
-    internal static void SetHandlerSettingsBasedOnOptions(this HttpClientHandler handler, Options options)
-    {
-        handler.AllowAutoRedirect = options.FollowRedirects;
-        handler.UseCookies = options.AutomaticCookieHandling;
+    using System;
+    using System.Net.Http;
+    using Frends.LinkedIn.SearchAdCampaigns.Definitions;
 
-        if (options.AllowInvalidCertificate)
-            handler.ServerCertificateCustomValidationCallback = (a, b, c, d) => true;
-    }
-
-    internal static void SetDefaultRequestHeadersBasedOnOptions(this HttpClient httpClient, Options options)
+    internal static class Extensions
     {
-        httpClient.DefaultRequestHeaders.ExpectContinue = false;
-        httpClient.DefaultRequestHeaders.TryAddWithoutValidation("content-type", "application/json");
-        httpClient.DefaultRequestHeaders.TryAddWithoutValidation("LinkedIn-Version", options.LinkedInAPIVersion);
-        httpClient.DefaultRequestHeaders.TryAddWithoutValidation("X-Restli-Protocol-Version", "2.0.0");
-        httpClient.Timeout = TimeSpan.FromSeconds(Convert.ToDouble(options.ConnectionTimeoutSeconds));
+        internal static void SetHandlerSettingsBasedOnOptions(this HttpClientHandler handler, Options options)
+        {
+            handler.AllowAutoRedirect = options.FollowRedirects;
+            handler.UseCookies = options.AutomaticCookieHandling;
+
+            if (options.AllowInvalidCertificate)
+                handler.ServerCertificateCustomValidationCallback = (a, b, c, d) => true;
+        }
+
+        internal static void SetDefaultRequestHeadersBasedOnOptions(this HttpClient httpClient, Options options)
+        {
+            httpClient.DefaultRequestHeaders.ExpectContinue = false;
+            httpClient.DefaultRequestHeaders.TryAddWithoutValidation("content-type", "application/json");
+            httpClient.DefaultRequestHeaders.TryAddWithoutValidation("LinkedIn-Version", options.LinkedInAPIVersion);
+            httpClient.DefaultRequestHeaders.TryAddWithoutValidation("X-Restli-Protocol-Version", "2.0.0");
+            httpClient.Timeout = TimeSpan.FromSeconds(Convert.ToDouble(options.ConnectionTimeoutSeconds));
+        }
     }
 }
